# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateCurrencyRequest:
    properties:
      asset_id:
        description: Optional Blockradar asset ID for crypto withdrawals (UUID-like).
        type: string
      coingecko_id:
        description: Optional CoinGecko coin ID for price fetching (e.g., "tether",
          "ethereum").
        type: string
      decimals:
        type: integer
      is_native:
        type: boolean
      name:
        type: string
      symbol:
        type: string
    required:
    - decimals
    - name
    - symbol
    type: object
  handler.CreateNetworkRequest:
    properties:
      chain_id:
        type: integer
      chain_key:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      network_type:
        enum:
        - mainnet
        - testnet
        type: string
    required:
    - chain_id
    - chain_key
    - name
    - network_type
    type: object
...
  models.Currency:
    properties:
      asset_id:
        description: |-
          AssetID is the Blockradar asset identifier used for withdrawals (UUID-like string).
          It is optional because not all currencies need to be withdrawable via Blockradar.
        type: string
      coingecko_id:
        description: |-
          CoinGeckoID is the CoinGecko coin identifier used for price fetching (e.g., "tether", "ethereum").
          It is optional and only needed for currencies that support relative pricing in P2P ads.
        type: string
      created_at:
        type: string
      decimals:
        type: integer
      id:
        type: string
      is_native:
        description: true for ETH
        type: boolean
      name:
        description: Ether, Tether
        type: string
      symbol:
        description: ETH, USDT
        type: string
      type:
        description: |-
          Type distinguishes fiat currencies (NGN, USD, etc.) from crypto assets.
          Values: "fiat" or "crypto". Defaults to "crypto" for backward compatibility.
          Existing fiat currency rows must be updated via a DB migration after deployment.
        type: string
      updated_at:
        type: string
    type: object
...
  models.Network:
    properties:
      chain_id:
        type: integer
      chain_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      network_type:
        description: mainnet | testnet
        type: string
      updated_at:
        type: string
    type: object
...
  models.PlatformFeeBalance:
    properties:
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      total_amount:
        type: number
      total_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  repository.DailyFeeRow:
    properties:
      currency:
        type: string
      date:
        description: '"YYYY-MM-DD"'
        type: string
      total_fees:
        type: number
      trade_count:
        type: integer
    type: object
...
  repository.OrderStats:
    properties:
      cancelled_orders:
        type: integer
      completed_orders:
        type: integer
      disputed_orders:
        type: integer
      total_orders:
        type: integer
      total_volume:
        description: sum of completed order totals (fiat)
        type: number
    type: object
...
  service.RevenueSummary:
    properties:
      fee_balances:
        items:
          $ref: '#/definitions/models.PlatformFeeBalance'
        type: array
      order_stats:
        $ref: '#/definitions/repository.OrderStats'
    type: object
...
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/utils.PaginationMeta'
      success:
        type: boolean
    type: object
  utils.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready backend API for Kynettic platform with authentication,
    user management, and health monitoring
  title: Kynettic API
  version: "1.0"
paths:
...
  /admin/currencies:
    post:
      consumes:
      - application/json
      description: 'Admin-only: create or update a currency (upsert by symbol).'
      parameters:
      - description: Currency
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCurrencyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Currency'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create/update currency
      tags:
      - admin
...
  /admin/kyc:
    get:
      description: 'Admin-only: paginated KYC queue with optional filters by status
        and tier.'
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10, max 100)
        in: query
        name: page_size
        type: integer
      - description: 'Filter by status: pending, verified, rejected'
        in: query
        name: status
        type: string
      - description: 'Filter by tier: 0, 1, 2, 3'
        in: query
        name: tier
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: List KYC verifications (admin)
      tags:
      - admin
  /admin/kyc/{user_id}:
    get:
      description: 'Admin-only: full KYC record for a user including address, selfie
        confidence, liveness score.'
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get KYC verification detail (admin)
      tags:
      - admin
  /admin/networks:
    post:
      consumes:
      - application/json
      description: 'Admin-only: create or update a network (upsert by chain_key +
        network_type).'
      parameters:
      - description: Network
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateNetworkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Network'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create/update network
      tags:
      - admin
...
  /admin/revenue/fees:
    get:
      description: 'Admin-only: fee totals grouped by day and currency. Default last
        30 days.'
      parameters:
      - description: 'Start date YYYY-MM-DD (default: 30 days ago)'
        in: query
        name: from
        type: string
      - description: 'End date YYYY-MM-DD inclusive (default: today)'
        in: query
        name: to
        type: string
      - description: Filter by currency symbol (e.g. USDT)
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/repository.DailyFeeRow'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Daily fee breakdown (admin)
      tags:
      - admin
  /admin/revenue/summary:
    get:
      description: 'Admin-only: lifetime fee balances per currency and order stats
        for the given date range (default last 30 days).'
      parameters:
      - description: 'Start date YYYY-MM-DD (default: 30 days ago)'
        in: query
        name: from
        type: string
      - description: 'End date YYYY-MM-DD inclusive (default: today)'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.RevenueSummary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Revenue summary (admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"